+++ app/DataObjects/Cocktail/Ingredient.php
@@ -7,7 +7,7 @@ namespace Kami\Cocktail\DataObjects\Cocktail;
 class Ingredient
 {
     /**
-     * @param array<int> $substitutes
+     * @param array<Substitute> $substitutes
      */
     public function __construct(
         public readonly int $id,
@@ -18,6 +18,7 @@ class Ingredient
         public readonly bool $optional = false,
         public readonly array $substitutes = [],
         public readonly ?float $amountMax = null,
+        public readonly ?string $note = null
     ) {
     }
 }