+++ app/DataObjects/Cocktail/Substitute.php
@@ -9,6 +9,7 @@ class Substitute
     public function __construct(
         public readonly int $ingredientId,
         public readonly ?float $amount = null,
+        public readonly ?float $amountMax = null,
         public readonly ?string $units = null,
     ) {
     }